+++ packages/dart/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **BREAKING CHANGES**: `numRandom` has renamed to `numPick`
 - `getFileName`: Fix incorrect directory name with include dot character
+- Add `getCopyFileName` method
 
 ## 1.1.11 (2025-12-10)